v.1.2.0
Refactored Shader Generator and added From Min and From Max sliders in panel
