+++ SciBlend/ShaderGenerator/operators/__init__.py
@@ -1,0 +1,1 @@
+"""Operators subpackage for Shader Generator (colorramp and material)."""
