+++ SciBlend/ShaderGenerator/ui/__init__.py
@@ -1,0 +1,1 @@
+"""UI subpackage for Shader Generator (Blender panels)."""
