+++ SciBlend/ShaderGenerator/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utility subpackage for Shader Generator (colormaps, attributes)."""
